voice_changer_app/requirements.txt
elevenlabs
pyaudio
python-dotenv
numpy
requests
customtkinter
packaging
